+++ front/src/navigation/RootNavigator.js
@@ -6,11 +6,7 @@ import PostListScreen from "../screens/PostListScreen";
 import CommunityScreen from "../screens/CommunityScreen";
 import ChatScreen from "../screens/ChatScreen";
 import MyPageScreen from "../screens/MyPageScreen";
-
-const MapScreen =
-  Platform.OS === "web"
-    ? require("../screens/MapScreen.web").default
-    : require("../screens/MapScreen").default;
+import MapScreen from "../screens/MapScreen";
 import { colors, spacing } from "../theme";
 
 const Tab = createBottomTabNavigator();
